Move main.py to layouts folder
